+++ Drivers/windriver.c
@@ -1,16 +1,10 @@
-//
-// Created by Tarikan on 09.03.2021.
-//
-
 #include "../driver.h"
 #include "memoryapi.h"
-#include "../arena.h"
 #include <windows.h>
 
-static int page_size;
+static DWORD page_size;
 
 void *kernel_alloc(size_t size) {
-
 
     return VirtualAlloc(NULL, size, MEM_RESERVE | MEM_COMMIT, PAGE_READWRITE);
 }
@@ -19,7 +13,7 @@ void kernel_free(void *ptr) {
     VirtualFree(ptr, 0, MEM_RELEASE);
 }
 
-int get_page_size() {
+size_t get_page_size() {
     if (!page_size) {
         SYSTEM_INFO si;
         GetSystemInfo(&si);
